refactor(native): extract sidecar auth state
